refactor _retrieve_table
